requirements.txt
beaker-py
PyYAML